fix: move typing and locale config to locale folder layout
