+++ src/app/[locale]/layout.tsx
@@ -1,3 +1,23 @@
-import LayoutRoot from '../layout'
+import type { PropsWithChildren } from 'react'
+import { NextIntlClientProvider } from 'next-intl'
+import { getMessages, setRequestLocale } from 'next-intl/server'
 
-export default LayoutRoot
+import { LanguageSelectorProvider } from 'components/language-selector/provider'
+
+type LayoutProps = PropsWithChildren<{
+  params: Promise<{ locale: string }>
+}>
+
+export default async function LocaleLayout({ children, params }: LayoutProps) {
+  const { locale } = await params
+  setRequestLocale(locale)
+  const messages = await getMessages()
+
+  return (
+    <LanguageSelectorProvider>
+      <NextIntlClientProvider messages={messages}>
+        {children}
+      </NextIntlClientProvider>
+    </LanguageSelectorProvider>
+  )
+}
